pyproject.toml
[tool.poetry]
name = "telegram-test-assigment"
version = "0.1.0"
description = ""
authors = ["cnwv <115669384+cnwv@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Pyrogram = "^2.0.106"
alembic = "^1.12.1"
sqlalchemy = "^2.0.23"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.1"
tgcrypto = "^1.2.5"
config = {path = "./config/"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
